make <settings> a component
