--- html/settings/index.html
@@ -1,46 +1,0 @@
-<div ng-controller="SettingsCtrl">
-  <div class="ph-header ph-lg-margin-bottom">
-    <div class="container">
-      <div class="row">
-        <div class="col-sm-12">
-          <div class="media">
-            <div class="media-left">
-              <!--icon-->
-              <span class="fa-stack ph-icon">
-                <i class="fa fa-circle fa-stack-2x fa-inverse"></i>
-                <i class="fa fa-cogs fa-stack-1x"></i>
-              </span>
-            </div>
-            <div class="media-body ph-vertical-middle">
-              <!--heading-->
-              <h2 class="media-heading">Settings</h2>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <!--subnav-->
-    <div class="ph-header-submenu-container">
-      <div class="container">
-        <div class="row">
-          <div class="col-sm-12">
-            <subnav>
-              <ul class="nav nav-tabs ph-header-submenu pull-left">
-                <li subnav-target="settings.profile">
-                  <a><i class="fa fa-fw fa-user"></i> Account</a>
-                </li>
-                <!--li subnav-target="settings.site">
-                  <a><i class="fa fa-fw fa-cog"></i> Site preferences</a>
-                </li-->
-              </ul>
-            </subnav>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <!--content-->
-  <div class="container" ng-form="settingsForm">
-    <div app-view-segment="1"></div>
-  </div>
-</div>
